fix py

+++ llm-service/app/routers/index/data_source/__init__.py
@@ -363,7 +363,7 @@ class DataSourceController:
             set(map(lambda r: r.data.params.get("user_name", "unknown"), runs))
         )
         count_of_direct_interactions = 0
-        max_score_over_time: list[tuple[int, int]] = []
+        max_score_over_time: list[tuple[int, float]] = []
         input_word_count_over_time: list[tuple[int, int]] = []
         output_word_count_over_time: list[tuple[int, int]] = []
         for run in relevant_runs:
